Fixed tests and moved methods for Catan
Fixed Catan tests (don't have to be in the view), and moved some view methods to another file
